Remove extra logic
In a node with ancestors and a right subtree, the right subtree always
contains the next largest node (because the ancestors are either smaller
than the node or larger than the node AND any of its children). Thus the
right subtree is the first place to check for the next largest node.